working tf_classification_example
